Separate sourcify fetching and util funcs
